+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/model/Model.java
@@ -40,6 +40,7 @@ public class Model {
     private final boolean timeRelativeID;
     private final SQLDatabaseModelExtension sqlDBModelExtension;
     private final BanyanDBModelExtension banyanDBModelExtension;
+    private final ElasticSearchModelExtension elasticSearchModelExtension;
 
     public Model(final String name,
                  final List<ModelColumn> columns,
@@ -50,7 +51,8 @@ public class Model {
                  final Class<?> streamClass,
                  boolean timeRelativeID,
                  final SQLDatabaseModelExtension sqlDBModelExtension,
-                 final BanyanDBModelExtension banyanDBModelExtension) {
+                 final BanyanDBModelExtension banyanDBModelExtension,
+                 final ElasticSearchModelExtension elasticSearchModelExtension) {
         this.name = name;
         this.columns = columns;
         this.scopeId = scopeId;
@@ -62,5 +64,6 @@ public class Model {
         this.timeRelativeID = timeRelativeID;
         this.sqlDBModelExtension = sqlDBModelExtension;
         this.banyanDBModelExtension = banyanDBModelExtension;
+        this.elasticSearchModelExtension = elasticSearchModelExtension;
     }
 }