+++ oap-server/server-library/library-client/src/main/java/org/apache/skywalking/oap/server/library/client/elasticsearch/IndexRequestWrapper.java
@@ -28,10 +28,17 @@ public class IndexRequestWrapper implements InsertRequest {
 
     public IndexRequestWrapper(String index, String type, String id,
                                Map<String, ?> source) {
+        this(index, type, id, null, source);
+    }
+
+    public IndexRequestWrapper(String index, String type, String id,
+                               String routing,
+                               Map<String, ?> source) {
         request = IndexRequest.builder()
                               .index(index)
                               .type(type)
                               .id(id)
+                              .routing(routing)
                               .doc(source)
                               .build();
     }