+++ oap-server/server-library/library-elasticsearch-client/src/main/java/org/apache/skywalking/library/elasticsearch/requests/factory/v7plus/codec/V7IndexRequestSerializer.java
@@ -23,6 +23,7 @@ import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import java.io.IOException;
 import org.apache.skywalking.library.elasticsearch.requests.IndexRequest;
+import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 final class V7IndexRequestSerializer extends JsonSerializer<IndexRequest> {
     @Override
@@ -37,6 +38,9 @@ final class V7IndexRequestSerializer extends JsonSerializer<IndexRequest> {
             {
                 gen.writeStringField("_index", value.getIndex());
                 gen.writeStringField("_id", value.getId());
+                if (StringUtil.isNotBlank(value.getRouting())) {
+                    gen.writeStringField("routing", value.getRouting());
+                }
             }
             gen.writeEndObject();
         }