+++ oap-server/server-storage-plugin/storage-elasticsearch-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/elasticsearch/base/RecordEsDAO.java
@@ -43,6 +43,7 @@ public class RecordEsDAO extends EsDAO implements IRecordDAO {
         Map<String, Object> builder = IndexController.INSTANCE.appendTableColumn(model, toStorage.obtain());
         String modelName = TimeSeriesUtils.writeIndexName(model, record.getTimeBucket());
         String id = IndexController.INSTANCE.generateDocId(model, record.id().build());
-        return getClient().prepareInsert(modelName, id, builder);
+        String routing = RoutingUtils.getRouting(model, builder);
+        return getClient().prepareInsert(modelName, id, routing, builder);
     }
 }